task-1/requirements.txt
# MVP: локальный OCR и обработка документов
# Системные зависимости (не pip): Tesseract OCR, Poppler (для pdf2image)
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0
openpyxl>=3.1.2
python-barcode[images]>=0.15.1
